02-llm-essentials/senior_search_system/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sentence-transformers==3.4.1
torch==2.6.0
numpy==2.2.4
pydantic==2.11.3
faiss-cpu==1.10.0
pytest==8.3.5
httpx==0.28.1
pytest-asyncio==0.26.0